// client-awb-middleoffice/src/app/model/coursBBE.ts
export class CoursBBE  {
    id: number;
    achatClientele: number;
    venteClientele: number;
    achatClienteleCAL: number;
    venteClienteleCAL: number;
    midBAM: number;
    achatinterBAM: number;
    venteinterBAM: number;
    rachatinter: number;
    venteinter: number;
    rachatsousdel: number;
    libDevise: string;
    uniteDevise: number;
    date: string;
  }
  